+++ packages/app_intents_codegen/lib/src/generator/xcstrings_generator.dart
@@ -30,6 +30,9 @@ class XcstringsGenerator {
       if (intent.resultDialogTemplate != null) {
         strings.add(intent.resultDialogTemplate!);
       }
+      if (intent.resultDialogSupportingTemplate != null) {
+        strings.add(intent.resultDialogSupportingTemplate!);
+      }
       if (intent.parameterSummary != null) {
         strings.add(_toSwiftKeyPathPlaceholders(intent.parameterSummary!));
       }